Check for OpenGL before running meshviewer
Previously, the command would dispatch regardless of the OpenGL
status.  In cases where OpenGL was unavailable, the process would
continue, start the client, and then run into an OpenGL error.
End-users would sometimes not be informed of errors.

This change checks for OpenGL prior to dispatch, alerting the user of
any errors.